requirements.txt
# Server
Flask == 1.0.2
Flask-Login == 0.4.1
Flask-SocketIO == 3.3.2
Flask-SQLAlchemy == 2.3.2
Flask-WTF == 0.14.2
WTForms == 2.2.1

# Tools and utilities
Werkzeug == 0.15.2
SQLAlchemy-utils == 0.33.11
psycopg2-binary == 2.8.1

# Deployment
gevent == 1.4.0
gevent-websocket == 0.10.1
gunicorn == 19.9.0
